Remove unnecessary code from the selector archiveContents.js.
The removed code block is executed if a group id was not found in the flat list of all groups (allGroups). The approach is to loop through all groups and search the id in a list of subgroups. But this makes no sense because the groups are stored in a flat list (allGroups) and therefore all groups have already been searched and the list of subgroups (allGroups[i].groups) never exists.